pyproject.toml
[tool.poetry]
name = "data-validation-module"
version = "0.1.7"
description = "A tool with different functions for the data validation."
authors = ["Lorenzo Olivier <lorenzo.o@seqana.com>"]
license = "Proprietary"
readme = "README.md"
repository = "https://gitlab.com:cquest1/prototypes/data_validation_module"

[tool.poetry.urls]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
flake8 = "^3.9.2"
loguru = "^0.5.3"
loguru-caplog="^0.2.0"
geopandas = "^0.9.0"
numpy="^1.21.2"
openpyxl="^3.0.7"
xlrd="^2.0.1"
file-path-tools = {git = "https://gitlab.com/cquest1/tools/file_path_tools", rev = "0.2.6"}
geopandas-tools = {git = "https://gitlab.com/cquest1/tools/geopandas_tools", rev = "0.1.3"}
Rtree = "^0.9.7"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
coverage = {extras = ["toml"], version = "^5.5"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.12.1"
xdoctest = {extras = ["colors"], version = "^0.15.5"}
pre-commit = "^2.13.0"
flake8 = "^3.9.2"
loguru = "^0.5.3"
black = "^20.8b1"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.4.3"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.0"
darglint = "^1.8.0"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.0.1"

[tool.poetry.scripts]


[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["get_gee_data"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=88
ensure_newline_before_comments=true
profile="black"

[tool.mypy]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
